use id instead of label in configurable content
